requirements.txt
opencv-python
torch
ultralytics
twilio
python-dotenv
